fix linking problem of warp flow
